simplify Promise code based on feedback
Promises: Kindly note that when calling fetch("https://pokeapi.co/api/v2/pokemon/squirtle/"), a promise will be returned to indicate if the HTTP request was successful(resolved) or not(rejected). Therefore, you do not need to use "resolve" and "reject" because the promise will handle this either way based on the HTTP request. I therefore encourage you to simplify your code and to use the code structure similar to what is provided on page 6 in the task document. Simply, remove line 1 and replace lines 12 - 14 with lines 25 - 27. Thereafter remove the block of code from line 22 onwards. Based on the above changes, you can simplify the code to first fetch the data from the API, use the first .then() in the promise chain to parse the response as a json object (meaning to convert the response into a JSON object), then use the second .then() to help you to work with the JSON data. You can thereafter use the console.log function to return the specific information from the API. Lastly, implement error handling, you could use the .catch() method to catch any errors if the fetch fails